fix some docs in perf test. fix scala 2.8.1 compile issue related to adding ints in SystemConsumers. revert accidental reviewboardrc checkin.

+++ samza-core/src/main/scala/org/apache/samza/system/SystemConsumers.scala
@@ -264,7 +264,7 @@ class SystemConsumers(
    * systemFetchMapCache stays in sync with fetchMap.
    */
   private def updateFetchMap(systemStreamPartition: SystemStreamPartition, amount: Int = 1) {
-    val fetchSize = fetchMap.getOrElse(systemStreamPartition, new Integer(0)) + amount
+    val fetchSize = fetchMap.getOrElse(systemStreamPartition, new Integer(0)).intValue + amount
     val systemName = systemStreamPartition.getSystem
     var systemFetchMap = systemFetchMapCache.getOrElse(systemName, Map())
